TST: signal: fix bspline tests for integer and complex inputs
Previously, ax = asarray(x) without the dtype was causing two
sets of issues:

- Integer inputs were wrong because floats were downcast to ints, e.g.
  `0.5 / 2 ==> 0`. This is gh-14530
- Tests for complex-valued data were just codifying the status quo,
  which was just assert the implementation details of
  cubic/quadratic/bspline routines.
  Evaluating a b-spline on a complex-valued argument makes no sense,
  these things are only defined on the real axis.

Also note that when it is changed, the whole `scipy.signal` test
suite passes. So this has no effect on the rest of the library.

+++ scipy/signal/_bsplines.py
@@ -184,7 +184,7 @@ def bspline(x, n):
     True
 
     """
-    ax = -abs(asarray(x))
+    ax = -abs(asarray(x, dtype=float))
     # number of pieces on the left-side is (n+1)/2
     funclist, condfuncs = _bspline_piecefunctions(n)
     condlist = [func(ax) for func in condfuncs]
@@ -337,7 +337,7 @@ def quadratic(x):
     True
 
     """
-    ax = abs(asarray(x))
+    ax = abs(asarray(x, dtype=float))
     res = zeros_like(ax)
     cond1 = less(ax, 0.5)
     if cond1.any():
